+++ ala-storm/pom.xml
@@ -19,7 +19,7 @@
             <groupId>org.apache.storm</groupId>
             <artifactId>storm-core</artifactId>
             <version>0.10.0</version>
-            <!--<scope>provided</scope>-->
+            <scope>compile</scope>
             <exclusions>
                 <exclusion>
                     <groupId>org.slf4j</groupId>
@@ -77,7 +77,7 @@
                                         implementation="org.apache.maven.plugins.shade.resource.ServicesResourceTransformer" />
                                 <transformer
                                         implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                                    <mainClass>com.unbxd.monitoring.alaStorm.Topology</mainClass>
+                                    <mainClass>com.unbxd.monitoring.alaStorm.TopologySubmit</mainClass>
                                 </transformer>
                             </transformers>
                         </configuration>
