+++ ala-storm/src/main/java/com/unbxd/monitoring/alaStorm/Topology.java
@@ -1,7 +1,6 @@
 package com.unbxd.monitoring.alaStorm;
 
 import backtype.storm.Config;
-import backtype.storm.LocalCluster;
 import backtype.storm.topology.TopologyBuilder;
 import backtype.storm.utils.Utils;
 import com.unbxd.monitoring.alaStorm.bolt.BoltFactory;
@@ -22,7 +21,7 @@ public class Topology {
     public Properties properties;
     public BoltFactory boltBuilder;
     public SpoutFactory spoutBuilder;
-    Config config;
+    public Config config;
 
 
     public Topology(String configFile, Config conf) throws Exception {
@@ -69,27 +68,4 @@ public class Topology {
 
     }
 
-    public static void main(String[] args) throws Exception {
-        String configFile = null;
-        if (args.length == 0) {
-            System.out.println("Missing input : config file location, using default");
-
-        } else {
-            configFile = args[0];
-        }
-
-        Config conf = new Config();
-        //conf.setDebug(true);
-        conf.setNumWorkers(2);
-        Topology ingestionTopology = new Topology(configFile, conf);
-
-        TopologyBuilder builder = ingestionTopology.buildTopology();
-        //StormSubmitter.submitTopology(topologyName, conf, builder.createTopology());
-        LocalCluster cluster = new LocalCluster();
-        cluster.submitTopology("test", conf, builder.createTopology());
-        Utils.sleep(1000*1000*1000);
-        cluster.killTopology("test");
-        cluster.shutdown();
-
-    }
 }
